Move sidebar state + events types to types module
- decluttering the main logic module
