+++ src/sidebar/annotations-sidebar/containers/logic.ts
@@ -1,199 +1,27 @@
 import debounce from 'lodash/debounce'
-import { UILogic, UIEvent, UIEventHandler, UIMutation } from 'ui-logic-core'
-import { TaskState } from 'ui-logic-core/lib/types'
+import { UILogic, UIEventHandler, UIMutation } from 'ui-logic-core'
 import { isFullUrl } from '@worldbrain/memex-url-utils'
 import { EventEmitter } from 'events'
 
 import { Annotation, AnnotationPrivacyLevels } from 'src/annotations/types'
-import { Anchor } from 'src/highlighting/types'
 import { loadInitial, executeUITask } from 'src/util/ui-logic'
-import { SidebarContainerDependencies } from './types'
+import type {
+    SidebarContainerDependencies,
+    SidebarContainerState,
+    SidebarContainerEvents,
+    EditForm,
+    EditForms,
+} from './types'
 import { AnnotationsSidebarInPageEventEmitter } from '../types'
-import { AnnotationMode } from 'src/sidebar/annotations-sidebar/types'
 import { DEF_RESULT_LIMIT } from '../constants'
-import { IncomingAnnotationData } from 'src/in-page-ui/shared-state/types'
 import {
     generateUrl,
     getLastSharedAnnotationTimestamp,
     setLastSharedAnnotationTimestamp,
 } from 'src/annotations/utils'
-import {
-    AnnotationSharingInfo,
-    AnnotationSharingAccess,
-} from 'src/content-sharing/ui/types'
 import { areTagsEquivalent } from 'src/tags/utils'
 import { FocusableComponent } from 'src/annotations/components/types'
-import { AnnotationsSorter } from '../sorting'
 import { CachedAnnotation } from 'src/annotations/annotations-cache'
-
-export interface EditForm {
-    isBookmarked: boolean
-    isTagInputActive: boolean
-    commentText: string
-    tags: string[]
-}
-
-export interface EditForms {
-    [annotationUrl: string]: EditForm
-}
-
-export interface SidebarContainerState {
-    loadState: TaskState
-    primarySearchState: TaskState
-    secondarySearchState: TaskState
-
-    showState: 'visible' | 'hidden'
-    isLocked: boolean
-
-    annotationSharingAccess: AnnotationSharingAccess
-    annotationSharingInfo: {
-        [annotationUrl: string]: AnnotationSharingInfo
-    }
-
-    showAllNotesCopyPaster: boolean
-    activeCopyPasterAnnotationId: string | undefined
-    activeTagPickerAnnotationId: string | undefined
-
-    pageUrl?: string
-    annotations: Annotation[]
-    annotationModes: {
-        [context in AnnotationEventContext]: {
-            [annotationUrl: string]: AnnotationMode
-        }
-    }
-    activeAnnotationUrl: string | null
-
-    showCommentBox: boolean
-    commentBox: EditForm
-
-    editForms: EditForms
-
-    pageCount: number
-    noResults: boolean
-
-    showCongratsMessage: boolean
-    showClearFiltersBtn: boolean
-    isSocialPost: boolean
-
-    // Filter sidebar props
-    showFiltersSidebar: boolean
-    showSocialSearch: boolean
-
-    annotCount?: number
-
-    // Search result props
-    shouldShowCount: boolean
-    isInvalidSearch: boolean
-    totalResultCount: number
-    allAnnotationsExpanded: boolean
-    searchResultSkip: number
-
-    isListFilterActive: boolean
-    isSocialSearch: boolean
-    showLoginModal: boolean
-    showAnnotationsShareModal: boolean
-    showBetaFeatureNotifModal: boolean
-
-    showAllNotesShareMenu: boolean
-    activeShareMenuNoteId: string | undefined
-    immediatelyShareNotes: boolean
-}
-
-export type SidebarContainerEvents = UIEvent<{
-    show: null
-    hide: null
-    lock: null
-    unlock: null
-
-    sortAnnotations: { sortingFn: AnnotationsSorter }
-
-    // Adding a new page comment
-    addNewPageComment: { comment?: string; tags?: string[] }
-    setNewPageCommentAnchor: { anchor: Anchor }
-    changeNewPageCommentText: { comment: string }
-    cancelEdit: { annotationUrl: string }
-    changeEditCommentText: { annotationUrl: string; comment: string }
-    saveNewPageComment: { privacyLevel: AnnotationPrivacyLevels }
-    cancelNewPageComment: null
-    updateNewPageCommentTags: { tags: string[] }
-
-    setEditCommentTagPicker: { annotationUrl: string; active: boolean }
-
-    updateTagsForEdit: {
-        added?: string
-        deleted?: string
-        annotationUrl: string
-    }
-    updateListsForPageResult: { added?: string; deleted?: string; url: string }
-    deleteEditCommentTag: { tag: string; annotationUrl: string }
-
-    receiveSharingAccessChange: {
-        sharingAccess: AnnotationSharingAccess
-    }
-
-    // Annotation boxes
-    goToAnnotationInNewTab: {
-        context: AnnotationEventContext
-        annotationUrl: string
-    }
-    setActiveAnnotationUrl: { annotationUrl: string }
-    setAnnotationEditMode: {
-        context: AnnotationEventContext
-        annotationUrl: string
-    }
-    editAnnotation: {
-        context: AnnotationEventContext
-        annotationUrl: string
-    }
-    deleteAnnotation: {
-        context: AnnotationEventContext
-        annotationUrl: string
-    }
-    shareAnnotation: {
-        context: AnnotationEventContext
-        annotationUrl: string
-        mouseEvent: React.MouseEvent
-    }
-    switchAnnotationMode: {
-        context: AnnotationEventContext
-        annotationUrl: string
-        mode: AnnotationMode
-    }
-
-    copyNoteLink: { link: string }
-    copyPageLink: { link: string }
-
-    setPageUrl: { pageUrl: string }
-
-    // Search
-    paginateSearch: null
-    setAnnotationsExpanded: { value: boolean }
-    toggleAllAnnotationsFold: null
-    fetchSuggestedTags: null
-    fetchSuggestedDomains: null
-
-    updateAnnotationShareInfo: {
-        annotationUrl: string
-        info: Partial<AnnotationSharingInfo>
-    }
-    updateAllAnnotationsShareInfo: {
-        info: AnnotationSharingInfo
-    }
-
-    setLoginModalShown: { shown: boolean }
-    setAnnotationShareModalShown: { shown: boolean }
-    setBetaFeatureNotifModalShown: { shown: boolean }
-
-    setAllNotesCopyPasterShown: { shown: boolean }
-    setCopyPasterAnnotationId: { id: string }
-    setTagPickerAnnotationId: { id: string }
-    resetTagPickerAnnotationId: null
-    resetCopyPasterAnnotationId: null
-
-    setAllNotesShareMenuShown: { shown: boolean }
-    resetShareMenuNoteId: null
-}>
-export type AnnotationEventContext = 'pageAnnotations' | 'searchResults'
 
 export type SidebarContainerOptions = SidebarContainerDependencies & {
     events?: AnnotationsSidebarInPageEventEmitter
